Se terminaron todas las funciones del frontend

+++ src/app/products/index-products/index-products.component.ts
@@ -10,10 +10,14 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
 import { SweetAlert2Module } from '@sweetalert2/ngx-sweetalert2';
 import { MatTableModule } from '@angular/material/table';
+import { AuthorizedComponent } from "../../security/authorized/authorized.component";
+import { MatIconModule } from '@angular/material/icon';
+import Swal from 'sweetalert2';
+import { SecurityService } from '../../security/security.service';
 
 @Component({
   selector: 'app-index-products',
-  imports: [RouterLink, ListGenericComponent, MatButtonModule, MatInputModule, SweetAlert2Module, MatPaginatorModule, MatTableModule],
+  imports: [RouterLink, ListGenericComponent, MatButtonModule, MatInputModule, SweetAlert2Module, MatPaginatorModule, MatTableModule, MatIconModule, AuthorizedComponent],
   templateUrl: './index-products.component.html',
   styleUrl: './index-products.component.css'
 })
@@ -21,15 +25,29 @@ export class IndexProductsComponent implements OnInit{
 
   ngOnInit(): void {
     this.uploadRecords();
+    this.fillColumnsToShow();
   }
 
   private productService = inject(ProductService);
+  private securityService = inject(SecurityService);
 
-  columnsToShow = ['id', 'image', 'name', 'price', 'acciones'];
+  columnsToShow: string[] = [];
 
   products!: ProductDTO[];
   pagination: PaginationDTO = {page: 1, recordsPerPage: 5};
   quantityTotalRecords!: number;
+
+  hasPermission(permission: string): boolean{
+    return this.securityService.hasPermission(permission);
+  }
+
+  fillColumnsToShow(){
+    if(this.securityService.getFieldJWT('Administrador') || this.hasPermission('Editar Producto') || this.hasPermission('Eliminar Producto')){
+      this.columnsToShow = ['id', 'image', 'name', 'price', 'acciones'];
+    }else{
+      this.columnsToShow = ['id', 'image', 'name', 'price'];
+    }
+  }
 
   uploadRecords(){
     this.productService.get(this.pagination).subscribe((response: HttpResponse<ProductDTO[]>) => {
@@ -47,9 +65,24 @@ export class IndexProductsComponent implements OnInit{
   }
 
   delete(id: number){
-    this.productService.delete(id).subscribe(() => {
-      this.pagination.page = 1;
-      this.uploadRecords();
+
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "No podrás revertir esta acción",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#027702',
+      cancelButtonColor: '#D70505',
+      confirmButtonText: 'Eliminar'
+    }).then((result) => {
+
+      if(result.isConfirmed){
+        
+        this.productService.delete(id).subscribe(() => {
+          this.pagination.page = 1;
+          this.uploadRecords();
+        });
+      }
     });
   }
